Rename ui menu service
