+++ ui/main/src/app/services/style/global-style.service.ts
@@ -10,7 +10,7 @@
 import {BehaviorSubject, Observable} from 'rxjs';
 import {UserPreferencesService} from '../userPreferences/UserPreferencesService';
 import {ConfigService} from 'app/services/config/ConfigService';
-import {MenuService} from 'app/business/services/menu.service';
+import {MenuService} from '@ofServices/menu/MenuService';
 
 declare const opfabStyle: any;
